iot-gateway/requirement.txt
confluent-kafka==2.6.1
pyyaml==6.0.2
requests==2.32.3
